Move the callback module to api_extensions
